Stat system and health bar

+++ Source/Assets/Scripts/Stats/Stat.cs
@@ -6,22 +6,22 @@ using UnityEngine;
 [System.Serializable]
 public class Stat
 {
-	public int baseValue;
+	public float baseValue;
 
-	List<int> modifiers = new List<int>();
+	List<float> modifiers = new List<float>();
 
-	public int GetValue(){
+	public float GetValue() {
 		var finalValue = baseValue;
 		finalValue += modifiers.Sum();
 		return finalValue;
 	}
 
-	public void AddModifier(int modifier){
+	public void AddModifier(float modifier){
 		if(modifier != 0)
 			modifiers.Add(modifier);
 	}
 
-	public void RemoveModifier(int modifier){
+	public void RemoveModifier(float modifier){
 		if(modifier != 0)
 			modifiers.Remove(modifier);
 	}
